mcp-server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-api-python-client>=2.122.0
google-auth>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
mistralai>=0.4.2
aiofiles>=23.2.1
pydantic>=2.6.1